+++ src/include/LSP/Workspace.hpp
@@ -97,7 +97,8 @@ public:
     std::optional<std::string> getDocumentationForType(const Luau::TypeId ty);
     std::optional<std::string> getDocumentationForAutocompleteEntry(const std::string& name, const Luau::AutocompleteEntry& entry,
         const std::vector<Luau::AstNode*>& ancestry, const Luau::ModuleName& moduleName);
-    std::vector<Reference> findAllReferences(const Luau::TypeId ty, std::optional<Luau::Name> property = std::nullopt);
+    std::vector<Reference> findAllTableReferences(const Luau::TypeId ty, std::optional<Luau::Name> property = std::nullopt);
+    std::vector<Reference> findAllFunctionReferences(const Luau::TypeId ty);
     std::vector<Reference> findAllTypeReferences(const Luau::ModuleName& moduleName, const Luau::Name& typeName);
 
     std::vector<lsp::CompletionItem> completion(const lsp::CompletionParams& params);